requirements.txt
Django==4.0.2
django_ratelimit2==0.4.2
django_recaptcha==3.0.0
python-dotenv==0.19.2
ratelimit==2.2.1
unicode_slugify==0.1.5
gunicorn==20.1.0
psycopg2-binary==2.9.3
